refactor(agent):rename ReportService
